+++ 2019/math/point.go
@@ -2,12 +2,12 @@ package math
 
 // Point represents a point in 2D space
 type Point struct {
-	X int32
-	Y int32
+	X int64
+	Y int64
 }
 
 // NewPoint creates a point with provided coordinates
-func NewPoint(x, y int32) Point {
+func NewPoint(x, y int64) Point {
 	return Point{
 		X: x,
 		Y: y,
@@ -15,6 +15,6 @@ func NewPoint(x, y int32) Point {
 }
 
 // ManhattanDistance calculates Manhattan distance between the points
-func (p Point) ManhattanDistance(other Point) int32 {
+func (p Point) ManhattanDistance(other Point) int64 {
 	return Abs(p.X-other.X) + Abs(p.Y-other.Y)
 }
